Make the indexing intrinsics fully Int32.

+++ src/native/intrinsics.jl
@@ -2,10 +2,7 @@
 
 export
     # Indexing and dimensions
-    # TODO: dynamically export
-    threadIdx,
-    blockDim, blockIdx,
-    gridDim,
+    threadIdx, blockDim, blockIdx, gridDim,
     warpsize,
 
     # Memory management
@@ -27,7 +24,7 @@ for dim in (:x, :y, :z)
             ($("""declare i32 @$intrinsic() readnone nounwind"""),
              $("""%1 = tail call i32 @$intrinsic()
                   ret i32 %1""")),
-            Int32, Tuple{}) + 1
+            Int32, Tuple{}) + Int32(1)
     end
 
     # Block size (#threads per block)
@@ -49,7 +46,7 @@ for dim in (:x, :y, :z)
             ($("""declare i32 @$intrinsic() readnone nounwind"""),
              $("""%1 = tail call i32 @$intrinsic()
                   ret i32 %1""")),
-            Int32, Tuple{}) + 1
+            Int32, Tuple{}) + Int32(1)
     end
 
     # Grid size (#blocks per grid)
@@ -94,7 +91,7 @@ sync_threads() = Base.llvmcall(
 # Shared memory
 #
 
-# FIXME: this is broken, cannot declare in `llvmcall`
+# FIXME: this is broken, cannot declare global variables in `llvmcall`
 # TODO: this is nasty
 #       - box-like semantics?
 #       - Ptr{AS}?
@@ -229,6 +226,7 @@ log10(x::Float64) = Base.llvmcall(
      """%2 = call double @__nv_log10(double %0)
         ret double %2"""),
     Float64, Tuple{Float64}, x)
+
 erf(x::Float32) = Base.llvmcall(
     ("""declare float @__nv_erff(float)""",
      """%2 = call float @__nv_erff(float %0)
